opens a new result/clicked result in a new nested tab
It adds a new tab at the front. The current results tabs aren't  stored on the server so they disappear on refresh. (Also can't close it yet)

sidenote I shortened the computation time
