+++ src/store.js
@@ -1,4 +1,11 @@
 // store.js
 import { reactive } from 'vue'
 
-export const store = reactive({ currentResult: [], queue: [] })
+const store = reactive({ currentResults: [], queue: [] })
+
+function addResult(result) {
+  store.currentResults = [result, ...store.currentResults]
+  console.log(store.currentResults)
+}
+
+export { store, addResult }
